backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytesseract
Pillow
reportlab
python-dotenv
requests
